+++ src/test/java/com/mobble/mobbleserver/support/fixture/article/ArticleTestFixture.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.support.fixture.article;
 
 import com.mobble.mobbleserver.refactor.article.entity.Article;
 import com.mobble.mobbleserver.refactor.article.entity.ArticleType;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.support.fixture.club.ClubTestFixture;
