+++ src/test/java/com/mobble/mobbleserver/support/fixture/chat/clubChatRoom/ClubChatRoomTestFixture.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.support.fixture.chat.clubChatRoom;
 
 import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoom;
 import com.mobble.mobbleserver.refactor.chat.clubChatRoom.entity.ClubChatRoom;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.support.fixture.chat.chatRoom.ChatRoomTestFixture;
 
 public final class ClubChatRoomTestFixture {
